Fixed Bug
Fixed a bug where when viewing a profile using a new database with the debug users that had no events it would crash the website because it was trying to display None events. I changed the html on line 98 in event-details.html so that if events None, it would not display anything.